deflate_medium: Split strategy into two, for levels 3-4 and 5-6. 3-4 does not use fizzle_matches, so the code becomes a lot simpler. 5-6 stays pretty much the same, but saves a few if checks.
